Jack </3: fix strange problem with modal event listener
Sometimes the event listener for the send form button was not being
added. it turned out that when isOwner or isCreatePage changed, the
button was being removed and then readded to the page and losing the
event listener.

I had to pull the links into a separate template, and add leanModal to the
onRendered method of that template. that way when isOwner changes
onRender of _navigationLinks is always called