Jack </3: having some strage issue...
where the modal popup for the tracking learn more link does not work
inside sandstorm (and only when running in sandstorm shell) - but the
Privacy link in the footer is fine :/

linking externally (with target=_blank) to work around this for now. but
it would be nice to have the Privacy modal pop up here too!